+++ packages/upgrader_core/lib/src/upgrade_state.dart
@@ -7,6 +7,7 @@ import 'upgrader_version_info.dart';
 
 class UpgraderState {
   UpgraderState({
+    this.checkOnResume = true,
     required this.client,
     this.clientHeaders,
     this.countryCodeOverride,
@@ -22,6 +23,9 @@ class UpgraderState {
     this.versionInfo,
   });
 
+  /// When `true`, the version info is retrieved from the store each time the
+  /// app is resumed from the background.
+  final bool checkOnResume;
   final http.Client client;
   final Map<String, String>? clientHeaders;
   final String? countryCodeOverride;
@@ -37,6 +41,7 @@ class UpgraderState {
   final UpgraderVersionInfo? versionInfo;
 
   UpgraderState copyWith({
+    bool? checkOnResume,
     http.Client? client,
     Map<String, String>? clientHeaders,
     String? countryCodeOverride,
@@ -52,6 +57,7 @@ class UpgraderState {
     UpgraderVersionInfo? versionInfo,
   }) {
     return UpgraderState(
+      checkOnResume: checkOnResume ?? this.checkOnResume,
       client: client ?? this.client,
       clientHeaders: clientHeaders ?? this.clientHeaders,
       countryCodeOverride: countryCodeOverride ?? this.countryCodeOverride,
@@ -78,6 +84,7 @@ class UpgraderState {
     bool? versionInfo,
   }) {
     return UpgraderState(
+      checkOnResume: checkOnResume,
       client: client,
       clientHeaders: clientHeaders,
       countryCodeOverride: